refactor(route): 统一路由处理架构，移除RootRoute结构
- 移除 RootRoute 结构体，将其功能完全融合到 Route 中
- 统一路由匹配逻辑，消除根路由和普通路由的人工区分
- 通过 configs 字段自然识别服务入口点，无需额外标志位
- 简化路由处理流程，减少代码重复和条件分支
- 保持所有原有功能和API兼容性
- 优化中间件层级管理和执行顺序
- 更新所有测试和示例以使用统一的Route结构

BREAKING CHANGE: RootRoute 类型已被移除，使用 Route::new_root() 替代
