+++ examples/middleware_layered/Cargo.toml
@@ -3,6 +3,8 @@ name = "middleware_layered"
 version = "0.1.0"
 edition = "2021"
 
+[workspace]
+
 [dependencies]
 silent = { path = "../../silent" }
 tokio = { version = "1.0", features = ["full"] }
