+++ examples/middleware_layered/src/main.rs
@@ -1,6 +1,6 @@
 use silent::prelude::*;
-use std::sync::Arc;
 use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::Arc;
 
 // 计数器中间件，用于演示中间件执行顺序
 #[derive(Clone)]
@@ -22,7 +22,11 @@ impl CounterMiddleware {
 impl MiddleWareHandler for CounterMiddleware {
     async fn handle(&self, req: Request, next: &Next) -> silent::Result<Response> {
         let count = self.counter.fetch_add(1, Ordering::SeqCst);
-        println!("🔧 {} middleware executed (count: {})", self.name, count + 1);
+        println!(
+            "🔧 {} middleware executed (count: {})",
+            self.name,
+            count + 1
+        );
 
         let response = next.call(req).await?;
         println!("🔧 {} middleware finished", self.name);
@@ -30,28 +34,28 @@ impl MiddleWareHandler for CounterMiddleware {
     }
 }
 
-async fn hello(_req: Request) -> Result<String, SilentError> {
+async fn hello(_req: Request) -> silent::Result<String> {
     println!("📍 Handler executed: hello");
     Ok("Hello from /api/v1/hello".to_string())
 }
 
-async fn world(_req: Request) -> Result<String, SilentError> {
+async fn world(_req: Request) -> silent::Result<String> {
     println!("📍 Handler executed: world");
     Ok("World from /api/v1/world".to_string())
 }
 
-async fn user_handler(_req: Request) -> Result<String, SilentError> {
+async fn user_handler(_req: Request) -> silent::Result<String> {
     println!("📍 Handler executed: user");
     Ok("User handler".to_string())
 }
 
-async fn root_handler(_req: Request) -> Result<String, SilentError> {
+async fn root_handler(_req: Request) -> silent::Result<String> {
     println!("📍 Handler executed: root");
     Ok("Root page".to_string())
 }
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
+async fn main() -> std::result::Result<(), Box<dyn std::error::Error>> {
     let counter = Arc::new(AtomicUsize::new(0));
 
     // 创建不同层级的中间件
@@ -62,22 +66,22 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     // 构建路由结构，每一层都有自己的中间件
     let app = Route::new("")
-        .hook(root_middleware)  // 根级中间件
+        .hook(root_middleware) // 根级中间件
         .get(root_handler)
         .append(
             Route::new("api")
-                .hook(api_middleware)  // API级中间件
+                .hook(api_middleware) // API级中间件
                 .append(
                     Route::new("v1")
-                        .hook(v1_middleware)  // V1级中间件
+                        .hook(v1_middleware) // V1级中间件
                         .get(hello)
                         .post(world)
                         .append(
                             Route::new("users")
-                                .hook(users_middleware)  // Users级中间件
-                                .get(user_handler)
-                        )
-                )
+                                .hook(users_middleware) // Users级中间件
+                                .get(user_handler),
+                        ),
+                ),
         );
 
     println!("🚀 启动层级中间件演示服务器...");
@@ -86,12 +90,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     println!("   GET  /api/v1/hello    - 应该执行: ROOT -> API -> V1 middleware");
     println!("   POST /api/v1/world    - 应该执行: ROOT -> API -> V1 middleware");
     println!("   GET  /api/v1/users    - 应该执行: ROOT -> API -> V1 -> USERS middleware");
-    println!("");
     println!("💡 每个路由层级独立管理自己的中间件");
     println!("💡 匹配到路由后，会按层级顺序执行所有相关中间件");
-    println!("");
 
-    let mut root_route = RootRoute::new();
+    let mut root_route = Route::new_root();
     root_route.push(app);
 
     let addr = "127.0.0.1:3000".parse()?;
