+++ silent/src/prelude.rs
@@ -25,7 +25,7 @@ pub use crate::handler::static_handler;
 pub use crate::log::*;
 pub use crate::middleware::MiddleWareHandler;
 pub use crate::route::handler_append::{HandlerAppend, HandlerGetter};
-pub use crate::route::{RootRoute, Route, RouteService, RouterAdapt};
+pub use crate::route::{Route, RouteService, RouterAdapt};
 #[cfg(feature = "scheduler")]
 pub use crate::scheduler::{SCHEDULER, SchedulerExt, Task};
 #[cfg(feature = "security")]
