+++ silent/src/route/mod.rs
@@ -1,6 +1,6 @@
 use async_trait::async_trait;
 use http::StatusCode;
-pub use root::RootRoute;
+// RootRoute 已被 Route 替代，不再导出
 pub use route_service::RouteService;
 use std::collections::HashMap;
 use std::fmt;
@@ -12,12 +12,57 @@ use crate::handler::static_handler;
 use crate::middleware::MiddleWareHandler;
 #[cfg(feature = "static")]
 use crate::prelude::HandlerGetter;
-use crate::{Method, Next, Request, Response, SilentError};
+use crate::route::handler_match::{Match, RouteMatched};
+use crate::{HandlerWrapper, Method, Next, Request, Response, SilentError};
 
 pub(crate) mod handler_append;
 mod handler_match;
-mod root;
 mod route_service;
+
+// LayeredHandler 从 root.rs 移过来
+struct LayeredHandler {
+    inner: RouteMatched,
+    middleware_layers: Vec<Vec<Arc<dyn MiddleWareHandler>>>,
+}
+
+#[async_trait]
+impl Handler for LayeredHandler {
+    async fn call(&self, req: Request) -> Result<Response, SilentError> {
+        match self.inner.clone() {
+            RouteMatched::Matched(route) => {
+                // 将所有层级的中间件扁平化，按顺序执行
+                let mut flattened_middlewares = vec![];
+                for layer in &self.middleware_layers {
+                    for middleware in layer {
+                        // 检查中间件是否匹配当前请求
+                        if middleware.match_req(&req).await {
+                            flattened_middlewares.push(middleware.clone());
+                        }
+                    }
+                }
+
+                let next = Next::build(Arc::new(route), flattened_middlewares);
+                next.call(req).await
+            }
+            RouteMatched::Unmatched => {
+                let handler = |_req| async move { Err::<(), SilentError>(SilentError::NotFound) };
+
+                // 对于未匹配的路由，仍然执行根级中间件（如果需要的话）
+                let mut root_middlewares = vec![];
+                if let Some(first_layer) = self.middleware_layers.first() {
+                    for middleware in first_layer {
+                        if middleware.match_req(&req).await {
+                            root_middlewares.push(middleware.clone());
+                        }
+                    }
+                }
+
+                let next = Next::build(Arc::new(HandlerWrapper::new(handler)), root_middlewares);
+                next.call(req).await
+            }
+        }
+    }
+}
 
 pub trait RouterAdapt {
     fn into_router(self) -> Route;
@@ -31,6 +76,10 @@ pub struct Route {
     pub middlewares: Vec<Arc<dyn MiddleWareHandler>>,
     special_match: bool,
     create_path: String,
+    // 配置管理字段（有此字段表示是服务入口点）
+    configs: Option<crate::Configs>,
+    #[cfg(feature = "session")]
+    session_set: bool,
 }
 
 impl RouterAdapt for Route {
@@ -71,6 +120,22 @@ impl fmt::Debug for Route {
 }
 
 impl Route {
+    /// 创建服务入口路由（原根路由功能）
+    /// 通过设置 configs 字段来标识这是一个服务入口点
+    pub fn new_root() -> Self {
+        Route {
+            path: String::new(),
+            handler: HashMap::new(),
+            children: Vec::new(),
+            middlewares: Vec::new(),
+            special_match: false,
+            create_path: String::new(),
+            configs: Some(crate::Configs::new()), // 服务入口点需要配置管理
+            #[cfg(feature = "session")]
+            session_set: false,
+        }
+    }
+
     pub fn new(path: &str) -> Self {
         let path = path.trim_start_matches('/');
         let mut paths = path.splitn(2, '/');
@@ -83,6 +148,9 @@ impl Route {
             middlewares: Vec::new(),
             special_match: first_path.starts_with('<') && first_path.ends_with('>'),
             create_path: path.to_string(),
+            configs: None,
+            #[cfg(feature = "session")]
+            session_set: false,
         };
         if last_path.is_empty() {
             route
@@ -137,11 +205,105 @@ impl Route {
     pub fn with_static_in_url(self, url: &str, path: &str) -> Self {
         self.append(Route::new(url).with_static(path))
     }
+
+    /// 添加子路由（原 RootRoute::push 功能）
+    pub fn push(&mut self, route: Route) {
+        self.children.push(route);
+    }
+
+    /// 添加中间件到当前路由首位（原 RootRoute::hook_first 功能）
+    pub fn hook_first(&mut self, handler: impl MiddleWareHandler + 'static) {
+        let handler = Arc::new(handler);
+        self.middlewares.insert(0, handler);
+    }
+
+    /// 设置配置（任何路由都可以使用）
+    pub fn set_configs(&mut self, configs: Option<crate::Configs>) {
+        self.configs = configs;
+    }
+
+    /// 获取配置
+    pub fn get_configs(&self) -> Option<&crate::Configs> {
+        self.configs.as_ref()
+    }
+
+    #[cfg(feature = "session")]
+    pub fn set_session_store<S: async_session::SessionStore>(&mut self, session: S) -> &mut Self {
+        self.hook_first(crate::session::middleware::SessionMiddleware::new(session));
+        self.session_set = true;
+        self
+    }
+
+    #[cfg(feature = "session")]
+    pub fn check_session(&mut self) {
+        if !self.session_set {
+            self.hook_first(crate::session::middleware::SessionMiddleware::default())
+        }
+    }
+
+    #[cfg(feature = "cookie")]
+    pub fn check_cookie(&mut self) {
+        self.hook_first(crate::cookie::middleware::CookieMiddleware::new())
+    }
+
+    #[cfg(feature = "template")]
+    pub fn set_template_dir(&mut self, dir: impl Into<String>) -> &mut Self {
+        let handler = crate::templates::TemplateMiddleware::new(dir.into().as_str());
+        self.middlewares.push(Arc::new(handler));
+        self
+    }
+
+    /// 作为服务入口点处理请求（包含路径匹配和中间件层级管理）
+    async fn handle_as_service_entry(
+        &self,
+        mut req: Request,
+    ) -> crate::error::SilentResult<Response> {
+        tracing::debug!("{:?}", req);
+        let configs = self.configs.clone().unwrap_or_default();
+        req.configs = configs.clone();
+
+        let (mut req, path) = req.split_url();
+
+        // 使用新的中间件收集逻辑
+        let (matched_route, middleware_layers) =
+            self.handler_match_collect_middlewares(&mut req, &path);
+
+        // 收集根级中间件
+        let mut root_middlewares = vec![];
+        for middleware in self.middlewares.iter().cloned() {
+            if middleware.match_req(&req).await {
+                root_middlewares.push(middleware);
+            }
+        }
+
+        // 将根级中间件添加到第一层
+        let mut all_middleware_layers = vec![];
+        if !root_middlewares.is_empty() {
+            all_middleware_layers.push(root_middlewares);
+        }
+        all_middleware_layers.extend(middleware_layers);
+
+        let handler = LayeredHandler {
+            inner: matched_route,
+            middleware_layers: all_middleware_layers,
+        };
+
+        // 直接调用 LayeredHandler
+        handler.call(req).await
+    }
 }
 
 #[async_trait]
 impl Handler for Route {
     async fn call(&self, req: Request) -> crate::error::SilentResult<Response> {
+        // 统一的路由处理逻辑
+
+        // 如果当前路由有配置，说明是服务入口点，需要处理路径匹配和中间件层级
+        if self.configs.is_some() {
+            return self.handle_as_service_entry(req).await;
+        }
+
+        // 普通路由的直接处理逻辑
         let configs = req.configs();
         match self.handler.get(req.method()) {
             None => Err(SilentError::business_error(
