+++ silent/src/route/route_service.rs
@@ -1,19 +1,19 @@
-use crate::route::{RootRoute, Route};
+use crate::route::Route;
 
 pub trait RouteService {
-    fn route(self) -> RootRoute;
-}
-
-impl RouteService for RootRoute {
-    fn route(self) -> RootRoute {
-        Self { ..self }
-    }
+    fn route(self) -> Route;
 }
 
 impl RouteService for Route {
-    fn route(self) -> RootRoute {
-        let mut root = RootRoute::new();
-        root.push(self);
-        root
+    fn route(self) -> Route {
+        // 如果已经是服务入口点（有配置），直接返回
+        if self.configs.is_some() {
+            self
+        } else {
+            // 否则创建新的根路由并添加当前路由为子路由
+            let mut root = Route::new_root();
+            root.push(self);
+            root
+        }
     }
 }
