+++ silent/src/service/serve.rs
@@ -1,18 +1,18 @@
 use crate::core::connection::Connection;
 use crate::core::socket_addr::SocketAddr;
-use crate::route::RootRoute;
+use crate::route::Route;
 use crate::service::hyper_service::HyperServiceHandler;
 use hyper_util::rt::{TokioExecutor, TokioIo};
 use hyper_util::server::conn::auto::Builder;
 use std::error::Error as StdError;
 
 pub(crate) struct Serve<E = TokioExecutor> {
-    pub(crate) routes: RootRoute,
+    pub(crate) routes: Route,
     pub(crate) builder: Builder<E>,
 }
 
 impl Serve {
-    pub(crate) fn new(routes: RootRoute) -> Self {
+    pub(crate) fn new(routes: Route) -> Self {
         Self {
             routes,
             builder: Builder::new(TokioExecutor::new()),
